backend/requirements.txt
alembic==1.4.2
appnope==0.1.0
backcall==0.1.0
click==7.1.2
decorator==4.4.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jupyter-client==6.1.3
jupyter-core==4.6.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.6.1
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pyzmq==19.0.0
six==1.14.0
SQLAlchemy==1.3.16
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
Werkzeug==1.0.1
